Rename bro to zeek: file names
