+++ libvast/src/system/default_application.cpp
@@ -16,11 +16,11 @@
 #include "vast/detail/assert.hpp"
 #include "vast/format/ascii.hpp"
 #include "vast/format/bgpdump.hpp"
-#include "vast/format/bro.hpp"
 #include "vast/format/csv.hpp"
 #include "vast/format/json.hpp"
 #include "vast/format/mrt.hpp"
 #include "vast/format/test.hpp"
+#include "vast/format/zeek.hpp"
 #include "vast/system/application.hpp"
 #include "vast/system/configuration.hpp"
 #include "vast/system/generator_command.hpp"
