+++ libvast/src/system/spawn_sink.cpp
@@ -23,9 +23,9 @@
 #include "vast/detail/make_io_stream.hpp"
 #include "vast/detail/unbox_var.hpp"
 #include "vast/format/ascii.hpp"
-#include "vast/format/bro.hpp"
 #include "vast/format/csv.hpp"
 #include "vast/format/json.hpp"
+#include "vast/format/zeek.hpp"
 #include "vast/system/node.hpp"
 #include "vast/system/sink.hpp"
 #include "vast/system/spawn_arguments.hpp"
