+++ libvast/test/system/source.cpp
@@ -21,7 +21,7 @@
 #include "vast/test/fixtures/actor_system_and_events.hpp"
 
 #include "vast/detail/make_io_stream.hpp"
-#include "vast/format/bro.hpp"
+#include "vast/format/zeek.hpp"
 #include "vast/system/atoms.hpp"
 #include "vast/table_slice.hpp"
 #include "vast/subset.hpp"
