+++ libvast_test/src/events.cpp
@@ -16,8 +16,8 @@
 #include "vast/default_table_slice_builder.hpp"
 #include "vast/defaults.hpp"
 #include "vast/format/bgpdump.hpp"
-#include "vast/format/bro.hpp"
 #include "vast/format/test.hpp"
+#include "vast/format/zeek.hpp"
 #include "vast/table_slice_builder.hpp"
 #include "vast/table_slice_builder_factory.hpp"
 #include "vast/table_slice_factory.hpp"
